+++ Sources/KeyboardKit/Layout/KeyboardLayout+Item.swift
@@ -13,8 +13,8 @@ public extension KeyboardLayout {
     
     /// A keyboard layout items defines an action, size, and
     /// optional insets for a key on a layout-based keyboard.
-    struct Item: Equatable {
-        
+    struct Item: Equatable, Sendable {
+
         /// Create a new layout item.
         ///
         /// - Parameters:
@@ -52,6 +52,29 @@ public extension KeyboardLayout {
 
     /// This typealias represents a list of layout item rows.
     typealias ItemRows = [ItemRow]
+}
+
+public extension KeyboardLayout.ItemRow {
+
+    /// Get a leading character margin action for the row, if any
+    var leadingCharacterMarginAction: KeyboardAction {
+        characterMarginAction(for: first)
+    }
+
+    /// Get a trailing character margin action for the row, if any
+    var trailingCharacterMarginAction: KeyboardAction {
+        characterMarginAction(for: last)
+    }
+
+    /// Get a matching character margin action for an action.
+    func characterMarginAction(
+        for item: KeyboardLayout.Item?
+    ) -> KeyboardAction {
+        switch item?.action {
+        case .character(let char): .characterMargin(char)
+        default: .none
+        }
+    }
 }
 
 public extension KeyboardLayout.Item {
@@ -98,8 +121,8 @@ public extension KeyboardLayout.Item {
 public extension KeyboardLayout {
     
     /// A size with point-based height and declarative width.
-    struct ItemSize: Equatable {
-        
+    struct ItemSize: KeyboardModel {
+
         /// Create a new layout item size.
         public init(
             width: ItemWidth,
@@ -117,8 +140,8 @@ public extension KeyboardLayout {
     }
     
     /// This enum specifies various layout item width types.
-    enum ItemWidth: Equatable {
-        
+    enum ItemWidth: KeyboardModel {
+
         /// A width that will share the available row space.
         case available